+++ R/00_global_variables.R
@@ -3,6 +3,6 @@ globalVariables(
         "run","rnorm","x","ds","date_col","value_col","n","as_tibble","value",
         "index","p50","diff_a","diff_b","ma12","year","month.lbl","wday.lbl",
         "week_of_month","week_day","fit","fft","time","x.y","y_actual","y_hat",
-        "dff_trans"
+        "dff_trans","sma_order","sma_value"
     )
 )