+++ NEWS.md
@@ -6,7 +6,8 @@ None
 ## New Features
 1. Fix #90 - Add `tidy_fft()` function
 2. Fix #92 - Add `ts_info_tbl()` function
-3. Fix #98 - Add `ts_to_tbl()` function
+3. Fix #96 - Add `ts_sma_plot()` function
+4. Fix #98 - Add `ts_to_tbl()` function
 
 ## Minor Fixes and Improvements
 None